// src/features/cross-chain/calculation-manager/providers/xy-provider/models/xy-cross-chain-quote-request.ts
export interface XyCrossChainQuoteRequest {
    srcChainId: number;
    srcQuoteTokenAddress: string;
    srcQuoteTokenAmount: string;
    dstChainId: number;
    dstQuoteTokenAddress: string;
    slippage: number;
    affiliate?: string;
    commissionRate?: number;
    srcSwapProvider?: string;
    bridgeProvider?: string;
    dstSwapProvider?: string;
}